// src/api/model/userProfilesPost.ts
/**
 * Financial Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserProfilesPost { 
    name: string;
    surname: string;
    email: string;
    password: string;
    taxIdentificationNumber: string;
    birthDate?: string;
    birthCountry?: string;
    birthState?: string;
    birthCity?: string;
}